Convert production files to float16 precision for memory optimization
- Updated pipeline scripts to save embeddings, whitening matrices, and quality grid as float16
- Reduced memory footprint: quality_scores_grid.npy from 237.72 MB to 59.43 MB (75% reduction)
- Reduced steam_tag_vectors.npy from 151.38 MB to 37.85 MB (75% reduction)
- All unit tests pass successfully
- Maintained backward compatibility with existing data structures
- Updated tasklist.md to mark task as completed

+++ pipeline/generate_quality_scores_grid.py
@@ -119,7 +119,7 @@ def generate_quality_grid(metadata_path, output_path):
         print(f"  Top 5 scores: {print_grid}")
 
     print(f"Saving grid to {output_path}...")
-    np.save(output_path, grid)
+    np.save(output_path, grid.astype(np.float16))
     print("Done!")
 
 if __name__ == "__main__":
